Some documentation of contest_link_80C2020
Merge itemsjson and mapjson into single tool
