+++ Makefile
@@ -25,8 +25,8 @@ PREPROC   := tools/preproc/preproc$(EXE)
 SCANINC   := tools/scaninc/scaninc$(EXE)
 RAMSCRGEN := tools/ramscrgen/ramscrgen$(EXE)
 GBAFIX    := tools/gbafix/gbafix$(EXE)
-ITEMSJSON := tools/itemsjson/itemsjson$(EXE)
-MAPJSON   := tools/mapjson/mapjson$(EXE)
+ITEMSJSON := tools/json2s/json2s$(EXE) items
+MAPJSON   := tools/json2s/json2s$(EXE) map
 
 ASFLAGS  := -mcpu=arm7tdmi -I include --defsym $(GAME_VERSION)=1 --defsym REVISION=$(GAME_REVISION) --defsym $(GAME_LANGUAGE)=1 --defsym DEBUG=$(DEBUG)
 CC1FLAGS := -mthumb-interwork -Wimplicit -Wparentheses -Wunused -Werror -O2 -fhex-asm
@@ -122,8 +122,7 @@ clean: tidy
 	$(MAKE) clean -C tools/aif2pcm
 	$(MAKE) clean -C tools/ramscrgen
 	$(MAKE) clean -C tools/gbafix
-	$(MAKE) clean -C tools/mapjson
-	$(MAKE) clean -C tools/itemsjson
+	$(MAKE) clean -C tools/json2s
 
 tools:
 	@$(MAKE) -C tools/gbagfx
@@ -135,8 +134,7 @@ tools:
 	@$(MAKE) -C tools/ramscrgen
 	@$(MAKE) -C tools/mid2agb
 	@$(MAKE) -C tools/gbafix
-	@$(MAKE) -C tools/mapjson
-	@$(MAKE) -C tools/itemsjson
+	@$(MAKE) -C tools/json2s
 
 tidy:
 	$(RM) $(ALL_BUILDS:%=poke%{.gba,.elf,.map})
