+++ data/specials.inc
@@ -84,20 +84,20 @@ gSpecials::
 	def_special CheckForBigMovieOrEmergencyNewsOnTV
 	def_special GetMomOrDadStringForTVMessage
 	def_special ResetTVShowState
-	def_special sub_80C47C0
-	def_special sub_80C496C
+	def_special ScrSpecial_GetContestWinnerIdx
+	def_special ScrSpecial_GetContestPlayerMonIdx
 	def_special sub_80C4C28
-	def_special sub_80C47F0
-	def_special sub_80C4858
+	def_special ScrSpecial_GetContestWinnerTrainerName
+	def_special ScrSpecial_GetContestWinnerNick
 	def_special sub_80C4C64
-	def_special sub_80C4758
+	def_special ScrSpecial_CountContestMonsWithBetterCondition
 	def_special SetContestTrainerGfxIds
-	def_special sub_80C43F4
+	def_special ScrSpecial_CheckSelectedMonAndInitContest
 	def_special sub_80C4D80
-	def_special sub_80C4F70
-	def_special sub_80C47A0
-	def_special sub_80C4440
-	def_special sub_80C44C0
+	def_special ScrSpecial_SetLinkContestTrainerGfxIdx
+	def_special ScrSpecial_GetMonCondition
+	def_special ScrSpecial_CanMonParticipateInSelectedLinkContest
+	def_special ScrSpecial_GiveContestRibbon
 	def_special sub_80C5044
 	def_special GiveMonArtistRibbon
 	def_special sub_808363C
