+++ include/contest.h
@@ -262,9 +262,9 @@ void LoadContestBgAfterMoveAnim(void);
 void CB2_StartContest(void);
 void Contest_CreatePlayerMon(u8);
 void Contest_InitAllPokemon(u8, u8);
-u8 sub_80AE47C(struct Pokemon *party);
-u16 sub_80AE770(u8, u8);
-void sub_80AE82C(u8);
+u8 CanMonParticipateInContest(struct Pokemon *party);
+u16 InitContestMonConditionI(u8, u8);
+void InitContestMonConditions(u8);
 u8 IsSpeciesNotUnown(u16);
 void sub_80AF668(void);
 void sub_80B0F28(u8);
