+++ include/contest_link_80C2020.h
@@ -2,12 +2,14 @@
 #define GUARD_CONTEST_LINK_80C2020_H
 
 void sub_80C2358(void);
-void sub_80C46EC(void);
-void sub_80C4740(void);
+void Contest_GetTrainerNameI_StringVar1(void);
+void Contest_GetNicknameI_StringVar1(void);
 void sub_80C48C8(void);
 void sub_80C48F4(void);
 void sub_80C4940(void);
 void sub_80C4980(u8);
 u8 sub_80C4B34(u8 *);
+void Contest_CopyAndConvertTrainerName_Intl(u8 *, const u8 *);
+void Contest_CopyAndConvertNicknameI_Intl(u8 *, u8);
 
 #endif // GUARD_CONTEST_LINK_80C2020_H
