--- tools/itemsjson/.gitignore
@@ -1,1 +1,0 @@
-itemsjson
