--- tools/itemsjson/itemsjson.h
@@ -1,29 +1,0 @@
-#ifndef METRONOME_SAPPHIRE_ITEMSJSON_H
-#define METRONOME_SAPPHIRE_ITEMSJSON_H
-
-#include <cstdio>
-using std::fprintf; using std::exit;
-
-#include <cstdlib>
-
-#ifdef _MSC_VER
-
-#define FATAL_ERROR(format, ...)          \
-do                                        \
-{                                         \
-    fprintf(stderr, format, __VA_ARGS__); \
-    exit(1);                              \
-} while (0)
-
-#else
-
-#define FATAL_ERROR(format, ...)            \
-do                                          \
-{                                           \
-    fprintf(stderr, format, ##__VA_ARGS__); \
-    exit(1);                                \
-} while (0)
-
-#endif // _MSC_VER
-
-#endif //METRONOME_SAPPHIRE_ITEMSJSON_H
