+++ tools/json2s/.gitignore
@@ -1,0 +1,2 @@
+json2s
+json2s.exe
