+++ tools/json2s/itemsjson.h
@@ -1,0 +1,8 @@
+#ifndef METRONOME_SAPPHIRE_ITEMSJSON_H
+#define METRONOME_SAPPHIRE_ITEMSJSON_H
+
+namespace itemsjson {
+    int main(int argc, char ** argv);
+}
+
+#endif //METRONOME_SAPPHIRE_ITEMSJSON_H
