+++ tools/json2s/mapjson.h
@@ -1,0 +1,8 @@
+#ifndef MAPJSON_H
+#define MAPJSON_H
+
+namespace mapjson {
+    int main(int argc, char ** argv);
+}
+
+#endif // MAPJSON_H
