--- tools/mapjson/.gitignore
@@ -1,1 +1,0 @@
-mapjson
